react router added with dropdown menu
