+++ package.json
@@ -59,5 +59,8 @@
       "**/hardhat",
       "**/hardhat/**"
     ]
+  },
+  "dependencies": {
+    "react-router-dom": "^6.22.3"
   }
 }
