+++ packages/react-app/src/pages/SwapPage.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+function SwapPage() {
+  return (
+    <div>
+      <h1>Here you can swap</h1>
+      <p>This is the info page of our application.</p>
+    </div>
+  );
+}
+
+export default SwapPage;
