+++ packages/react-app/src/pages/WalletPage.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import MainWallet from "../MainWallet";
+
+function WalletPage(subgraphUri) {
+  return <MainWallet subgraphUri={subgraphUri} />;
+}
+
+export default WalletPage;
